requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
click==8.0.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-model-utils==4.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
gunicorn==20.1.0
Markdown==3.3.4
mypy-extensions==0.4.3
numpy==1.21.0
pandas==1.2.5
pathspec==0.8.1
Pillow==8.2.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
reportlab==3.5.68
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
